Remove API to create ane event consumer thread
