+++ src/main/java/net/rubygrapefruit/platform/internal/jni/AbstractFileEventFunctions.java
@@ -44,85 +44,6 @@ public abstract class AbstractFileEventFunctions implements NativeIntegration {
          * @see FileWatcher#startWatching(Collection)
          */
         public abstract FileWatcher start() throws InterruptedException;
-
-        /**
-         * Starts the file watcher with a thread consuming the events and
-         * feeding them to the given {@link FileWatchEvent.Handler}.
-         *
-         * The started thread will terminate once the returned
-         * {@link FileWatcher} is {@link FileWatcher#close() closed}.
-         */
-        public FileWatcher startWithHandler(FileWatchEvent.Handler handler) throws InterruptedException {
-            return startWithHandler(new ThreadConfigurator() {
-                @Override
-                public void configure(Thread processorThread) {
-                    // Do nothing
-                }
-            }, handler);
-        }
-
-        /**
-         * Starts the file watcher with a thread consuming the events and
-         * feeding them to the given {@link FileWatchEvent.Handler}.
-         *
-         * The started thread will terminate once the returned
-         * {@link FileWatcher} is {@link FileWatcher#close() closed},
-         * or when it is {@link Thread#interrupt() interrupted}.
-         *
-         * Allows the configuration of the thread created via the given
-         * {@link ThreadConfigurator} before it is started.
-         */
-        public FileWatcher startWithHandler(ThreadConfigurator threadConfigurator, final FileWatchEvent.Handler handler) throws InterruptedException {
-            Thread processorThread = new Thread(new Runnable() {
-                private volatile boolean terminated;
-
-                @Override
-                public void run() {
-                    while (!terminated && !Thread.interrupted()) {
-                        FileWatchEvent event;
-                        try {
-                            event = eventQueue.take();
-                        } catch (InterruptedException e) {
-                            break;
-                        }
-                        event.handleEvent(new FileWatchEvent.Handler() {
-                            @Override
-                            public void handleChangeEvent(FileWatchEvent.ChangeType type, String absolutePath) {
-                                handler.handleChangeEvent(type, absolutePath);
-                            }
-
-                            @Override
-                            public void handleUnknownEvent(String absolutePath) {
-                                handler.handleUnknownEvent(absolutePath);
-                            }
-
-                            @Override
-                            public void handleOverflow(FileWatchEvent.OverflowType type, String absolutePath) {
-                                handler.handleOverflow(type, absolutePath);
-                            }
-
-                            @Override
-                            public void handleFailure(Throwable failure) {
-                                handler.handleFailure(failure);
-                            }
-
-                            @Override
-                            public void handleTerminated(boolean successful) {
-                                terminated = true;
-                                handler.handleTerminated(successful);
-                            }
-                        });
-                    }
-                }
-            }, "File watcher event handler");
-            threadConfigurator.configure(processorThread);
-            processorThread.start();
-            return start();
-        }
-    }
-
-    interface ThreadConfigurator {
-        void configure(Thread processorThread);
     }
 
     /**
